+++ DAL/Entities/PTdPenPrepAttachementDtl.cs
@@ -13,7 +13,7 @@ public partial class PTdPenPrepAttachementDtl
 {
     [Key]
     [Column("INT_ATTACHEMENT_DTL_ID")]
-    public int IntAttachementDtlId { get; set; }
+    public long IntAttachementDtlId { get; set; }
 
     [Column("DOCUMENT_ID")]
     public int DocumentId { get; set; }
